# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/orders": {
      "get": {
        "description": "Filtrlar: holat, manba, tezkor, to‘lov holati, filial, menejer, sana oralig‘i, qidiruv (raqam / kompaniya / ism / telefon).\n\n🔒 Xodim faqat o‘z filialini ko‘radi; SUPER_ADMIN `branchId` bilan filtrlaydi.",
        "operationId": "OrdersAdminController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Saralanadigan maydon nomi",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "NEW",
                "SEARCHING_TRANSPORT",
                "LOADING",
                "DELIVERING",
                "DELIVERED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "source",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "WEBSITE",
                "TELEGRAM",
                "PHONE",
                "ADMIN"
              ]
            }
          },
          {
            "name": "isUrgent",
            "required": false,
            "in": "query",
            "description": "Faqat tezkor (`true`) yoki oddiy",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "paymentStatus",
            "required": false,
            "in": "query",
            "description": "Shu holatdagi to‘lovi bor buyurtmalar",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PAID",
                "FAILED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "branchId",
            "required": false,
            "in": "query",
            "description": "Filial filtri — SUPER_ADMIN uchun. Boshqa rol o‘zinikidan boshqasini bersa — 404.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "managerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Yaratilgan sana (UTC) — shu kundan boshlab, kiritiladi",
            "schema": {
              "example": "2026-09-01",
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Yaratilgan sana (UTC) — shu kungacha, KIRITILMAYDI",
            "schema": {
              "example": "2026-10-01",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Buyurtma raqami, kompaniya nomi, telefon yoki mijoz ismi",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sahifalangan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PaginatedAdminOrderListItemDtoDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Filtr noto‘g‘ri (masalan dateFrom ≥ dateTo)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Boshqa filial so‘raldi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buyurtmalar ro‘yxati",
        "tags": [
          "Admin"
        ]
      },
      "post": {
        "description": "Xaridor: `customerId` (hisobi bor mijoz) YOKI `guestName` + `guestPhone`.\n\n🔒 Narx backendda: mijozda — uning filiali narxi va shaxsiy qoidalari; hisobsiz xaridorda — filial bazaviy narxi. Zaxira tekshiruvi mijoz buyurtmasi bilan bir xil.\n\nMenejer kiritsa — buyurtma unga biriktiriladi.",
        "operationId": "OrdersAdminController_createManual",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManualOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Buyurtma yaratildi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminOrderDetailDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Xaridor noto‘g‘ri ko‘rsatilgan yoki buyurtma tarkibi xato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Mijoz/filial/mahsulot topilmadi yoki boshqa filialniki",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Omborda yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buyurtmani qo‘lda kiritish (telefon / Telegram)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/orders/{id}": {
      "get": {
        "description": "Tarkib, holatlar tarixi (kim o‘zgartirgani bilan), to‘lovlar.",
        "operationId": "OrdersAdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Buyurtma ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Buyurtma",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminOrderDetailDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Topilmadi yoki boshqa filialga tegishli",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buyurtma tafsiloti",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/orders/{id}/urgent": {
      "patch": {
        "operationId": "OrdersAdminController_setUrgent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Buyurtma ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetOrderUrgentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Yangilangan buyurtma",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminOrderDetailDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Topilmadi yoki boshqa filialga tegishli",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tezkor belgisini qo‘yish / olib tashlash",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/orders/{id}/assign": {
      "patch": {
        "description": "Xodim buyurtma filialining faol xodimi (MANAGER, BRANCH_ADMIN yoki MODERATOR) bo‘lishi shart. `managerId: null` — biriktirishni olib tashlash.",
        "operationId": "OrdersAdminController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Buyurtma ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignOrderManagerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Yangilangan buyurtma",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminOrderDetailDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Xodim topilmadi, faol emas yoki boshqa filialda",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Topilmadi yoki boshqa filialga tegishli",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buyurtmaga xodim biriktirish",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/orders/{id}/status": {
      "patch": {
        "description": "Yetkazib berish: `NEW → SEARCHING_TRANSPORT → LOADING → DELIVERING → DELIVERED`.\n\nOlib ketish: `NEW → LOADING → DELIVERED`.\n\n`CANCELLED` — yuk yo‘lga chiqquncha (NEW, SEARCHING_TRANSPORT, LOADING). `DELIVERED` va `CANCELLED` — yakuniy.\n\nHar o‘zgarish tarixga yoziladi (kim, qachon, izoh). Javobdagi `allowedNextStatuses` — keyingi tugmalar uchun.\n\n🔒 Faqat o‘z filiali buyurtmasi (SUPER_ADMIN — hammasi).",
        "operationId": "OrdersAdminController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Buyurtma ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Yangi holat va tarix",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/OrderStatusChangeResponseDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ruxsat etilmagan o‘tish yoki o‘sha holat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Topilmadi yoki boshqa filialga tegishli",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Holat bir vaqtda boshqa xodim tomonidan o‘zgartirildi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buyurtma holatini o‘zgartirish",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/v1/me/orders/{id}": {
      "get": {
        "description": "Tarkib (buyurtma paytidagi narx bilan), to‘lovlar, holatlar tarixi vaqt belgisi bilan.\n\n🔒 Faqat o‘z buyurtmasi — begonasi 404.",
        "operationId": "MeOrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Buyurtma ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Buyurtma",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/OrderCustomerResponseDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token yo‘q, muddati o‘tgan yoki hisob faol emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Optom mijoz tokeni emas yoki vaqtinchalik parol almashtirilmagan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buyurtma tafsiloti",
        "tags": [
          "Customers"
        ]
      }
    }
  },
  "info": {
    "title": "Vodiy Kafel API",
    "description": "Vodiy Kafel savdo platformasining backend API'si.\n\n### Javob formati\nBarcha muvaffaqiyatli javoblar `{ \"data\": ... }` ichida keladi.\nSahifalangan ro'yxatlarda qo'shimcha `meta` bo'ladi.\n\nBarcha xatolar bir xil formatda — `ApiErrorDto` modeliga qarang.\n\n### Autentifikatsiya\nToken `Authorization: Bearer <token>` sarlavhasida yuboriladi.\nYuqoridagi **Authorize** tugmasi orqali kiritishingiz mumkin.\n\n### Muhim qoidalar\n- Mahsulot javoblarida ombordagi **aniq son yo'q** — faqat `stockStatus`.\n- Buyurtma summasini backend o'zi hisoblaydi; frontend yuborgan narx e'tiborga olinmaydi.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status kodi"
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "Status nomi"
          },
          "message": {
            "description": "Xato matni. Validatsiya xatosida — matnlar massivi, aks holda bitta matn.",
            "oneOf": [
              {
                "type": "string",
                "example": "Mahsulot topilmadi"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "pallets must not be less than 1"
                ]
              }
            ]
          },
          "path": {
            "type": "string",
            "example": "/api/v1/products",
            "description": "So‘rov manzili"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-08T17:02:11.419Z",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "example": "1339b1a6-5cf7-4d0f-9d4b-10bfbd295f49",
            "description": "Kuzatuv ID. Xato haqida xabar berganda shu ID ni ayting — log‘dan aniq so‘rov topiladi."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp"
        ]
      },
...
      "QuoteItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "description": "Mahsulot ID"
          },
          "pallets": {
            "type": "number",
            "description": "Paddon soni (kv² emas — TZ 3.3)",
            "minimum": 1,
            "maximum": 100000,
            "example": 10
          }
        },
        "required": [
          "productId",
          "pallets"
        ]
      },
...
      "PaginatedAdminOrderListItemDtoDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminOrderListItemDto"
            }
          },
          "total": {
            "type": "number",
            "example": 128,
            "description": "Jami elementlar soni"
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "example": 7,
            "description": "Jami sahifalar soni"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
...
      "CreateManualOrderDto": {
        "type": "object",
        "properties": {
          "items": {
            "minItems": 1,
            "maxItems": 50,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteItemDto"
            }
          },
          "regionId": {
            "type": "string",
            "description": "Yetkazib berish viloyati — `transportTypeId` bilan BIRGA. Ikkalasi ham berilmasa — olib ketish (yo‘l kira 0)."
          },
          "transportTypeId": {
            "type": "string",
            "description": "Transport turi — `regionId` bilan BIRGA"
          },
          "exactLat": {
            "type": "number",
            "description": "Xaritada belgilangan aniq nuqta (TZ 3.13) — `exactLng` bilan birga, faqat yetkazib berishda. ⚠ Narxga TA’SIR QILMAYDI — faqat logistika.",
            "minimum": -90,
            "maximum": 90,
            "example": 41.311081
          },
          "exactLng": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "example": 69.240562
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "CASH",
              "CARD",
              "BANK_TRANSFER"
            ],
            "description": "`CASH` — naqd, `CARD` — karta orqali onlayn, `BANK_TRANSFER` — shartnoma bo‘yicha o‘tkazma"
          },
          "note": {
            "type": "string",
            "maxLength": 1000
          },
          "source": {
            "type": "string",
            "enum": [
              "PHONE",
              "TELEGRAM",
              "ADMIN"
            ],
            "example": "PHONE"
          },
          "customerId": {
            "type": "string",
            "description": "Hisobi bor optom mijoz. Berilsa — guestName/guestPhone yo‘q."
          },
          "guestName": {
            "type": "string",
            "maxLength": 150,
            "example": "Aziz aka"
          },
          "guestPhone": {
            "type": "string",
            "example": "+998901234567"
          },
          "branchId": {
            "type": "string",
            "description": "Hisobsiz xaridor uchun filial — faqat SUPER_ADMIN beradi. Xodim uchun har doim o‘z filiali; mijozda — mijozning filiali."
          },
          "isUrgent": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "items",
          "paymentMethod",
          "source"
        ]
      },
      "SetOrderUrgentDto": {
        "type": "object",
        "properties": {
          "isUrgent": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "isUrgent"
        ]
      },
      "AssignOrderManagerDto": {
        "type": "object",
        "properties": {
          "managerId": {
            "type": "string",
            "nullable": true,
            "description": "Shu buyurtma filialining faol xodimi. `null` — biriktirishni olib tashlash."
          }
        },
        "required": [
          "managerId"
        ]
      },
...
    }
  }
}